move spot light function into spot light file
